feat: html fix and occlude materials
