+++ src/core/UI/Html/HTML.stories.tsx
@@ -6,11 +6,15 @@ import { useTurntable } from '@sb/useTurntable'
 
 import { Icosahedron, Html, OrthographicCamera } from 'drei'
 import { HtmlProps, CalculatePosition } from './Html'
+import { HtmlMaterial as HtmlMaterialLegacy } from '@legacy/Materials/HtmlMaterial'
+import { HtmlMaterial as HtmlMaterialWebGPU } from '../../../webgpu/Materials/HtmlMaterial'
+import { PlatformSwitch } from '@sb/components/PlatformSwitch'
 import { useFrame, useThree } from '@react-three/fiber'
 import { Meta, StoryObj } from '@storybook/react-vite'
 
 export default {
   title: 'UI/Html',
+  tags: ['dual'],
   component: Html,
   decorators: [
     (Story, context) => (
@@ -172,21 +176,59 @@ function HTMLOccluderScene(props: HtmlProps) {
       <group ref={turntableRef}>
         <Icosahedron name="pink" args={[5, 5]} position={[0, 0, 0]}>
           <meshBasicMaterial color="hotpink" />
-          <Html {...props} position={[0, 0, -6]} className="html-story-label" occlude="blending">
-            Blending
-          </Html>
+          <PlatformSwitch
+            legacy={
+              <Html
+                {...props}
+                position={[0, 0, -6]}
+                className="html-story-label"
+                occlude="blending"
+                material={<HtmlMaterialLegacy transform={false} />}
+              >
+                Blending
+              </Html>
+            }
+            webgpu={
+              <Html
+                {...props}
+                position={[0, 0, -6]}
+                className="html-story-label"
+                occlude="blending"
+                material={<HtmlMaterialWebGPU transform={false} />}
+              >
+                Blending
+              </Html>
+            }
+          />
         </Icosahedron>
         <Icosahedron name="yellow" args={[5, 5]} position={[16, 0, 0]}>
           <meshBasicMaterial color="yellow" />
-          <Html
-            {...props}
-            transform
-            position={[0, 0, -6]}
-            className="html-story-label html-story-label-B"
-            occlude="blending"
-          >
-            Blending w/ transform
-          </Html>
+          <PlatformSwitch
+            legacy={
+              <Html
+                {...props}
+                transform
+                position={[0, 0, -6]}
+                className="html-story-label html-story-label-B"
+                occlude="blending"
+                material={<HtmlMaterialLegacy transform={true} />}
+              >
+                Blending w/ transform
+              </Html>
+            }
+            webgpu={
+              <Html
+                {...props}
+                transform
+                position={[0, 0, -6]}
+                className="html-story-label html-story-label-B"
+                occlude="blending"
+                material={<HtmlMaterialWebGPU transform={true} />}
+              >
+                Blending w/ transform
+              </Html>
+            }
+          />
         </Icosahedron>
         <Icosahedron ref={occluderRef} name="orange" args={[5, 5]} position={[0, 0, 16]}>
           <meshBasicMaterial color="orange" />
