+++ src/legacy/Materials/HtmlMaterial/index.ts
@@ -1,0 +1,1 @@
+export * from './HtmlMaterial'
