+++ src/legacy/Materials/index.ts
@@ -24,5 +24,6 @@ export * from './SoftShadows'
 export * from './SpotLightMaterial'
 export * from './WireframeMaterial'
 export * from './CloudMaterial'
+export * from './HtmlMaterial'
 export * from './StarsMaterial'
 export * from './SparklesMaterial'
