+++ src/webgpu/Materials/index.ts
@@ -20,5 +20,6 @@ export * from '../Effects/Outlines'
 export * from './SpotLightMaterial'
 export * from './WireframeMaterial'
 export * from './FakeCloudMaterial'
+export * from './HtmlMaterial'
 export * from './StarsMaterial'
 export * from './SparklesMaterial'
